Fix cpu speed/max speed reading

+++ Shared/src/Models/CPU/CpuUsage.cs
@@ -5,11 +5,11 @@ namespace Monitrix.System.Models;
 
 public sealed class CpuUsageModel
 {
-    public double CpuSpeedMhz { get; init; }
+    public float CpuSpeedMhz { get; init; }
 
-    public double CurrentSpeedMhz { get; init; }
+    public float CurrentSpeedMhz { get; init; }
 
-    public double TotalUsagePercentage { get; init; }
+    public float TotalUsagePercentage { get; init; }
 
     public IReadOnlyList<float> PerCoreUsagePercentage { get; init; } = Array.Empty<float>();
 }
